requirements.txt
kiteconnect==5.0.1
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0
schedule>=1.2.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
